Refactor towards a more typical project structure
Alter app setup and file structure slightly so that they're more
indicative of how a non-trivial Flask app might be structured. Specifically:

- Define the app factory in manage/__init__.py and keep basic app setup
  code there
- Make use of blueprints to organise routes and views
